requirements.txt
# Core dependencies (tested versions)
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.1.0
pyyaml==6.0.1

# ROS 2 dependencies
# rosbag2_py
# rclpy
# sensor_msgs
# nav_msgs
# tf2_ros
# cv_bridge

# For systems without ROS 2, we'll use sqlite3 (built-in) to read bags directly
# and implement custom message parsing

# Object detection
ultralytics==8.0.196  # YOLOv8
torch==2.1.0
torchvision==0.16.0

# Point cloud processing
open3d==0.17.0

# Utilities
tqdm==4.66.1
scipy==1.11.3
matplotlib==3.8.0
scikit-learn==1.3.2  # For DBSCAN clustering


